Assert Emacs version in Travis CI tests
The current Travis CI setup is very tolerant against failures to
install the requested Emacs version, falling back to the default
Emacs 24.3.1.  This can be very confusing when debugging errors
because the value for $EMACS_VERSION does not always correlate with
the actual Emacs version the tests were run against.

This change adds an assertion test that ensures that the actual Emacs
version the tests are running against is the one requested by
$EMACS_VERSION.  For emacs-snapshot, as there is no property "this is
a development snapshot", this change relies on emacs-version having
four elements (e. g., "26.0.50.1").